# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinanceFlow API",
        "description": "FinanceFlow API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server development (localhost)"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server development (IP)"
        }
    ],
    "paths": {
...
        "/api/v1/recurring-transactions": {
            "get": {
                "tags": [
                    "Recurring Transactions"
                ],
                "summary": "Get all recurring transactions for the authenticated user",
                "operationId": "e1f9a88190f772628099e7fe272d5214",
                "responses": {
                    "200": {
                        "description": "List of recurring transactions"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recurring Transactions"
                ],
                "summary": "Create a new recurring transaction template",
                "description": "Transfers are not supported for recurring transactions.",
                "operationId": "6a515d92854fef13adcc04dd39a08209",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "account_id",
                                    "type",
                                    "amount",
                                    "currency_code",
                                    "frequency",
                                    "next_run_at"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Netflix subscription"
                                    },
                                    "account_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 3,
                                        "nullable": true
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "expense",
                                        "enum": [
                                            "income",
                                            "expense"
                                        ]
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15.99
                                    },
                                    "currency_code": {
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "example": "monthly",
                                        "enum": [
                                            "daily",
                                            "weekly",
                                            "monthly",
                                            "yearly"
                                        ]
                                    },
                                    "next_run_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-01"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Streaming service",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recurring transaction created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/recurring-transactions/{id}": {
            "get": {
                "tags": [
                    "Recurring Transactions"
                ],
                "summary": "Get a specific recurring transaction",
                "operationId": "bc2e5589f5be382180892a9304ce6902",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurring transaction details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Recurring Transactions"
                ],
                "summary": "Delete a recurring transaction",
                "operationId": "92822db11df3bf921186b44e475efd7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Recurring transaction deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Recurring Transactions"
                ],
                "summary": "Update a recurring transaction",
                "operationId": "e49c978d4e45b39ac531f1a4da294dbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Subscription"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 19.99
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "example": "yearly",
                                        "enum": [
                                            "daily",
                                            "weekly",
                                            "monthly",
                                            "yearly"
                                        ]
                                    },
                                    "next_run_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2027-01-01"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recurring transaction updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}